backend/requirements.txt
# === Core API ===
fastapi
uvicorn
pydantic
python-dotenv

# === Data science ===
numpy
pandas
scipy

# === Database ===
sqlalchemy
psycopg2-binary

# === Geospatial ===
pyproj
shapely
fiona
geopandas
geojson

# === Utilities ===
requests
click
colorama
